Fix: Make things work on 32-bit when set by corflags
