+++ source/OpenBVE/OpenBve.csproj
@@ -195,10 +195,15 @@
     <Compile Include="System\Functions\Bve5Shared.cs" />
     <Compile Include="System\Functions\CrashHandler.cs" />
     <Compile Include="System\Functions\FolderBrowserDialog.cs" />
+    <Compile Include="System\Input\JoystickManager\AbstractJoystick.cs" />
+    <Compile Include="System\Input\JoystickManager\AbstractJoystickManager.cs" />
     <Compile Include="System\Input\Controls.Control.cs" />
     <Compile Include="System\Input\Controls.cs" />
     <Compile Include="System\Input\InputDevicePlugin.cs" />
-    
+    <Compile Include="System\Input\JoystickManager\JoystickManager32.cs" />
+    <Compile Include="System\Input\JoystickManager\JoystickManager64.cs" />
+    <Compile Include="System\Input\JoystickManager\RailDriver\AbstractRailDriver.cs" />
+    <Compile Include="System\Input\JoystickManager\StandardJoystick.cs" />
     <Compile Include="System\Logging\BlackBox.cs" />
     <Compile Include="System\Logging\Score.cs" />
     <Compile Include="System\Options.cs" />
@@ -287,19 +292,8 @@
     <Compile Include="OldCode\TrainManager.cs" />
     <Compile Include="OldCode\World.cs" />
     <Compile Include="System\Host.cs" />
-    <Compile Include="System\Input\JoystickManager.cs" />
-    <Compile Include="System\Input\JoystickManager.RailDriver.cs">
-      <DependentUpon>JoystickManager.cs</DependentUpon>
-    </Compile>
-    <Compile Include="System\Input\JoystickManager.StandardJoystick.cs">
-      <DependentUpon>JoystickManager.cs</DependentUpon>
-    </Compile>
-	<Compile Include="System\Input\JoystickManager.RailDriver32.cs">
-      <DependentUpon>JoystickManager.cs</DependentUpon>
-    </Compile>
-    <Compile Include="System\Input\JoystickManager.RailDriver64.cs">
-      <DependentUpon>JoystickManager.cs</DependentUpon>
-    </Compile>
+    <Compile Include="System\Input\JoystickManager\RailDriver\RailDriver32.cs" />
+    <Compile Include="System\Input\JoystickManager\RailDriver\RailDriver64.cs" />
     <Compile Include="System\Program.cs" />
     <EmbeddedResource Include="UserInterface\formBugReport.resx">
       <DependentUpon>formBugReport.cs</DependentUpon>
